can eat pieces now
Almost done!  I need to add special abilities to spotter, bomb, flag,
and scout then a system to show pieces when they are revealed.  After I
finish that it will be fully playable and all I will need to work on is
the AI which may or may not get done.